+++ Riot/Modules/Pills/PillTextAttachmentData.swift
@@ -1,23 +1,14 @@
 // 
-// Copyright 2022 New Vector Ltd
+// Copyright 2022-2024 New Vector Ltd.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: AGPL-3.0-only
+// Please see LICENSE in the repository root for full details.
 //
 
 import Foundation
 import UIKit
 
-@available (iOS 15.0, *)
+@available(iOS 15.0, *)
 struct PillAssetColor: Codable {
     var red: CGFloat = 0.0, green: CGFloat = 0.0, blue: CGFloat = 0.0, alpha: CGFloat = 0.0
     
@@ -30,7 +21,7 @@ struct PillAssetColor: Codable {
     }
 }
 
-@available (iOS 15.0, *)
+@available(iOS 15.0, *)
 struct PillAssetParameter: Codable {
     var tintColor: PillAssetColor?
     var backgroundColor: PillAssetColor?
@@ -38,7 +29,7 @@ struct PillAssetParameter: Codable {
     var padding: CGFloat = 2.0
 }
 
-@available (iOS 15.0, *)
+@available(iOS 15.0, *)
 enum PillTextAttachmentItem: Codable {
     case text(String)
     case avatar(url: String?, string: String?, matrixId: String)
@@ -46,7 +37,7 @@ enum PillTextAttachmentItem: Codable {
     case asset(named: String, parameters: PillAssetParameter)
 }
 
-@available (iOS 15.0, *)
+@available(iOS 15.0, *)
 extension PillTextAttachmentItem {
     var string: String? {
         switch self {
@@ -59,7 +50,7 @@ extension PillTextAttachmentItem {
 }
 
 /// Data associated with a Pill text attachment.
-@available (iOS 15.0, *)
+@available(iOS 15.0, *)
 struct PillTextAttachmentData: Codable {
     // MARK: - Properties
     /// Pill type