+++ Riot/Modules/Pills/PillsFormatter.swift
@@ -1,24 +1,15 @@
 // 
-// Copyright 2022 New Vector Ltd
+// Copyright 2022-2024 New Vector Ltd.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: AGPL-3.0-only
+// Please see LICENSE in the repository root for full details.
 //
 
 import Foundation
 import UIKit
 
 /// Provides utilities funcs to handle Pills inside attributed strings.
-@available (iOS 15.0, *)
+@available(iOS 15.0, *)
 @objcMembers
 class PillsFormatter: NSObject {
     // MARK: - Internal Properties
@@ -214,7 +205,7 @@ class PillsFormatter: NSObject {
 }
 
 // MARK: - Private Methods
-@available (iOS 15.0, *)
+@available(iOS 15.0, *)
 extension PillsFormatter {
     struct MarkdownLinkResult: Equatable {
         let url: URL